fix(desktop): setup items act, and a linear goal keeps its text
The setup checklist listed what was missing without letting anyone act on it.
An unmet item is now a control that lands on the surface which resolves it, and
the two steps with no honest destination stay inert and say what would satisfy
them instead of pretending.

A goal built from a Linear task was cut at 1200 characters, so the tail of the
description was silently lost.

While in the file, the onboarding card moved off function declarations and its
sub-components moved to sibling files, and the palette event name became one
shared constant instead of a literal repeated on both sides.

Closes #1424
Closes #1429
