+++ apps/desktop/src/features/onboarding/openCommandPaletteEvent.ts
@@ -1,0 +1,1 @@
+export const OPEN_COMMAND_PALETTE_EVENT = 'goodboy:open-command-palette';
